allow non-3-tuple keys when assisted sext decoding in ldb fold_object fun
Because flexible keys work by @andytill has lifted the 3-field
restriction on TS keys, we should, correspondingly, allow the 5th
byte of the binary (which is the number of elements in the encoded
tuple) to be any number.

With the tuple size hardcoded at 3, keys with the number of fields other
than 3 were treated as non-TS data, and were returned to the key fold
function already sext-decoded, eventually breaking the client code which
attempted decoding on its own. In particular, there was a workaround for
this, which is now being removed (the change in
riak_kv_wm_timeseries_listkeys:ts_keys_to_body).