convert iso tstamps in qry:do_insert, not in ts_util:build_sql_record_int
In order to convert timestamps given as strings, rather doing this in
riak_kv_ts_util:build_sql_record_int, we pass values for INSERT as is and
relegate the timestamp conversion to riak_kv_qry:do_insert. The conversion
will thus take place after checking for too many values wrt the number of
fields. Otherwise, this check would need to be done twice as the timestamp
conversion code needs to have the field types available (and check that
length(Types) == length(Values), and do something meaningful about it if
it is not the case).

This also ensures the too_many_values error is still diagnosed and reported
before any timestamp conversion errors.