index_fsm:init: avoid passing fun objects as args to remote call sites
Because instances of index fsm may be started on different nodes
(particularly, on different nodes with differing versions of the
beams), passing a fun object to where the call is made is likely to
cause a badfun error the moment the function object is called.  This
happens to work when the module beam on the remote node contains
the compiled code the received function object refers to; but all
chances are off when 'index' and/or 'uniq' properties are different
between the caller and target beams.

To avoid badfuns, we export the function we are passing as an
argument, for the caller to refer to it safely.

This won't fix badfuns in a mixed cluster where clients connect to 1.3
nodes to call API functions involving list_keys but will prevent this from
happening going forward.

See www.javalimit.com/2010/05/passing-funs-to-other-erlang-nodes.html for
details and discussion.