Make w1c conflict resolution predictable
By design, the write-once code path has deterministic
but unpredictable sibling resolution. The objects are
hashed and compared.

This was intended to discourage write-once updates.

Instead, choose the "last" write as determined by
comparing timestamps on the version vectors. This
can still fail to produce the desired results if
the clocks on servers are out of sync, but will
usually give the right answer.