export & use riak_kv_ts_util:local_to_partition_key as KeyConvFun
Previously, a KeyConvFun used to be constructed on coordinator and sent
to vnodes, where it was passed to backend and used in key filter.  This
works fine as long as vnode and coordinator nodes are of the same version.

This patch, instead, passes a {M, F} tuple, in order for the conversion
function to be constructed on-site (i.e., on the node where it is to be
executed), thus avoiding the dreaded badfun error.

This should allow list_keys to work in mixed-cluster scenarios going
forward from version 1.5.  Note that list_keys will still fail for 1.4+1.5
clusters.