single key delete
An implementation of single-key deletes of TS data via the standard
delete path.

For TS data, a compound key is required to generate partition key and
local key.  It is collected from elements in #tsdelreq{} (rather than
along to riak_client:delete as {PK, LK}.

In riak_kv_delete:delete, PK and LK are used to create a tombstone
object.

Complementary changes are in riak_pb (introducing pb messages tsdelreq
and tsdelresp with codes 94 and 95).

Note additionally that we
* don't collect all the quorum-related options from #tsdelreq
  (be consistent with existing put call);
* don't take fresh vclock by default for delete op: fetch an
  existing object's one instead, like it's done for regular deletes;
* take care to pick apart Key = {PK, LK} of TS objects in riak_kv_get_fsm:
  use PK for coverage calculation, LK as regular key for lookups;

# Conflicts:
#	src/riak_client.erl
#	src/riak_kv_app.erl
#	src/riak_kv_delete.erl
#	src/riak_kv_get_fsm.erl
#	src/riak_kv_qry_worker.erl