check table readiness for TS operations
If a table's bucket type is created but not activated (such as with
separate invocations of riak-admin), all TS functions will fail.  A
wrapper with a check for this condition is added in appropriate places
in various riak_kv_pb_timeseries:process/2.

An error code 1019 is returned when a TS function is called on a non-
activated bucket type.

# Conflicts:
#	src/riak_kv_pb_timeseries.erl