Refactor to use monitoring
Discussion of the reliability of process monitoring, including the monitoring of processes on remote nodes, led to a new design which simplifies the message passing required by relying instead on monitoring.

A request should now be refused if there exists an currently active grant, and the preflist has changed since that grant was activated.  In all other situations grants should be made (either immediately or after being queued awaiting a release).