Use token_manager to control downstream messages
The token_manager will no longer monitor remote upstream sessions.  Add a backup GC process, whereby if a request is blocked, there is a check of the existence in the remote manager of an association with the session.

The only thing that must be reliable, is therefore the monitoring of local sessions by managers.

Before using a local session, an association check is now made, that the local token manager has an active association for that session.