Avoid re-fetching objects
Once a HEAD has been converted to a GET, disqualify it from the find_bestobject calculation.

As GETs occur after HEADs in the GET_FSM, it should by default be the case when allow_mult=true that VCget >= VChead.

This is not the case with lww=true, wher eobjects may regress in VC terms.  However as lww=true, it is destined to prefer the GET which came after the HEAD.

The code as-is could cause a HEAD to continuously require re-fetching, and lead to requests timing out.  This is demonstrated in the riak_test associated with the conditional PUT token implementation